# apps/primary-backend/openapi.yaml
openapi: 3.0.0
info:
  title: Super-30 API
  version: 1.0.0
servers:
  - url: http://localhost:3100
paths:
  /api/v1/action/available:
    get:
      summary: Get available actions
      responses:
        "200":
          description: List of available actions
          content:
            application/json:
              schema:
                type: object
                properties:
                  availableActions:
                    type: array
                    items:
                      type: object
  /api/v1/credentials:
    get:
      summary: Get credentials for user
      security:
        - apiKeyAuth: []
      responses:
        "200":
          description: List of credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  credentials:
                    type: array
                    items:
                      type: object
    post:
      summary: Create credentials
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                platform:
                  type: string
                  enum: [email, telegram]
                keys:
                  type: object
      responses:
        "200":
          description: Credentials stored successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    delete:
      summary: Delete credentials
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                credentialsId:
                  type: string
      responses:
        "200":
          description: Credentials deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /api/v1/trigger/available:
    get:
      summary: Get available triggers
      responses:
        "200":
          description: List of available triggers
          content:
            application/json:
              schema:
                type: object
                properties:
                  availableTriggers:
                    type: array
                    items:
                      type: object
  /api/v1/user/signup:
    post:
      summary: Sign up a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                name:
                  type: string
      responses:
        "200":
          description: User signed up
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /api/v1/user/signin:
    post:
      summary: Sign in a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: User signed in
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /api/v1/workflow:
    post:
      summary: Create a workflow
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                availableTriggerId:
                  type: string
                title:
                  type: string
                actions:
                  type: array
                  items:
                    type: object
      responses:
        "200":
          description: Workflow created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "411":
          description: Invalid Inputs
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "400":
          description: Invalid action metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  errors:
                    type: array
                    items:
                      type: object
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    get:
      summary: Get workflows for user
      security:
        - apiKeyAuth: []
      responses:
        "200":
          description: List of workflows
          content:
            application/json:
              schema:
                type: object
                properties:
                  workflows:
                    type: array
                    items:
                      type: object
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /api/v1/workflow/{id}:
    get:
      summary: Get a workflow by id
      security:
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Workflow details
          content:
            application/json:
              schema:
                type: object
                properties:
                  workflow:
                    type: object
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    put:
      summary: Update a workflow
      security:
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                availableTriggerId:
                  type: string
                title:
                  type: string
                actions:
                  type: array
                  items:
                    type: object
      responses:
        "200":
          description: Workflow updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "411":
          description: Invalid Inputs
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "403":
          description: Workflow does not belong to user
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization